+++ back_end/auth/modules/auth/services/ClientService.js
@@ -1,12 +1,3 @@
-// exports.createClient = async (data) => {
-//     await DB_CONNECTION.models.Client.create(data)
-// }
-
-exports.updateClient = async (data) => {
-    // const update = await DB_CONNECTION.models.Client.create(data)
-    // await update.save()
-}
-
 exports.getClient = async (data) => {
     try {
         const client = await DB_CONNECTION.models.Client.findOne({ where: data })
